Django/Homeworks/Week9/requirements.txt
actionlib==1.13.2
angles==1.9.13
asgiref==3.4.1
backports.zoneinfo==0.2.1
bondpy==1.8.6
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
controller-manager==0.19.5
controller-manager-msgs==0.19.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cv-bridge==1.16.0
Deprecated==1.2.13
diagnostic-analysis==1.10.4
diagnostic-common-diagnostics==1.10.4
diagnostic-updater==1.10.4
Django==4.0
django-appconf==1.0.5
django-extensions==3.1.5
django-filter==21.1
django-oauth-toolkit==1.5.0
django-staticfiles==1.2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
drf-yasg2==1.19.4
dynamic-reconfigure==1.7.1
future==0.18.2
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
idna==3.3
image-geometry==1.16.0
importlib-metadata==4.9.0
inflection==0.5.1
interactive-markers==1.12.0
itypes==1.2.0
Jinja2==3.0.3
joint_state_publisher==1.15.0
joint_state_publisher_gui==1.15.0
jwcrypto==1.0
laser_geometry==1.6.7
Markdown==3.3.6
MarkupSafe==2.0.1
message-filters==1.15.13
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
pkg_resources==0.0.0
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.6
python-qt-binding==0.4.4
pytz==2021.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
requests==2.26.0
resource_retriever==1.12.6
rosbag==1.15.13
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.13
roslaunch==1.15.13
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.13
rosmake==1.15.8
rosmaster==1.15.13
rosmsg==1.15.13
rosnode==1.15.13
rosparam==1.15.13
rospy==1.15.13
rosservice==1.15.13
rostest==1.15.13
rostopic==1.15.13
rosunit==1.15.8
roswtf==1.15.13
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.13
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.2
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.2
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
rviz==1.14.11
sensor-msgs==1.13.1
serializers==0.2.4
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
sqlparse==0.4.2
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
topic-tools==1.15.13
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
xacro==1.14.10
zipp==3.6.0
